Move to input package
