+++ Games/EchoGame/src/main/java/jokatu/game/games/echo/Echo.java
@@ -1,6 +1,7 @@
 package jokatu.game.games.echo;
 
 import jokatu.game.event.PublicGameEvent;
+import jokatu.game.games.echo.input.EchoInput;
 import org.jetbrains.annotations.NotNull;
 
 public class Echo implements PublicGameEvent {
