+++ Games/EchoGame/src/main/java/jokatu/game/games/echo/EchoGameFactory.java
@@ -2,6 +2,7 @@ package jokatu.game.games.echo;
 
 import jokatu.game.AbstractGameFactory;
 import jokatu.game.GameID;
+import jokatu.game.games.echo.input.EchoInput;
 import jokatu.game.input.InputDeserialiser;
 import jokatu.game.player.PlayerFactory;
 import org.jetbrains.annotations.NotNull;
