+++ Games/EchoGame/src/main/java/jokatu/game/games/echo/input/EchoInputAcceptor.java
@@ -1,6 +1,8 @@
-package jokatu.game.games.echo;
+package jokatu.game.games.echo.input;
 
 import jokatu.game.event.AbstractPrivateGameEvent;
+import jokatu.game.games.echo.Echo;
+import jokatu.game.games.echo.EchoPlayer;
 import jokatu.game.input.InputAcceptor;
 import org.jetbrains.annotations.NotNull;
